requirements.txt
google-generativeai>=0.3.0
Django>=5.0.2
python-dotenv>=1.1.0
requests>=2.32.3